psycho/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.6
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.5
uvicorn==0.22.0
yarl==1.9.2